Better modularize blend crypto module
