+++ nomos-blend/network/src/core/tests/utils.rs
@@ -10,7 +10,11 @@ use libp2p::{
 };
 use libp2p_swarm_test::SwarmExt as _;
 use nomos_blend_message::{
-    crypto::{Ed25519PrivateKey, ProofOfQuota, ProofOfSelection, Signature, SIGNATURE_SIZE},
+    crypto::{
+        keys::Ed25519PrivateKey,
+        proofs::{quota::ProofOfQuota, selection::ProofOfSelection},
+        signatures::{Signature, SIGNATURE_SIZE},
+    },
     input::EncapsulationInput,
     PayloadType,
 };
