+++ nomos-services/blend/src/membership/mod.rs
@@ -4,7 +4,7 @@ pub mod service;
 use std::pin::Pin;
 
 use futures::Stream;
-use nomos_blend_message::crypto::Ed25519PublicKey;
+use nomos_blend_message::crypto::keys::Ed25519PublicKey;
 use nomos_blend_scheduling::membership::Membership;
 use overwatch::services::{relay::OutboundRelay, ServiceData};
