+++ nomos-services/blend/src/membership/service.rs
@@ -1,7 +1,7 @@
 use std::{collections::BTreeSet, hash::Hash, marker::PhantomData};
 
 use futures::StreamExt as _;
-use nomos_blend_message::crypto::Ed25519PublicKey;
+use nomos_blend_message::crypto::keys::Ed25519PublicKey;
 use nomos_blend_scheduling::membership::{Membership, Node};
 use nomos_core::sdp::{Locator, ProviderId, ServiceType};
 use nomos_membership::{
